Update font size when the screen changes
This is a very conservative change -- it only happens if the window
height matches perfectly and only on the glutin backend.

But switching to fullscreen and back on a 1920x1080 screen looks
great.

One issue I've noticed is: the scrolling in fullscreen (windowed is
fine) becomes very choppy. My laptop's been giving me troubles so this
needs further investigation. Maybe it's been choppy all the time.