Read tilesize from settings during recenter anim
This fixes the jumpiness during the animation when we centre the
screen.

It was using the `engine::TILESIZE` constant rather than the new
possibly changing value and so the animation calculation was off.

We should probably not use the constant anywhere, have default
settings which use it as the default value and just use the settings
value everywhere.