Make the TILESIZE const i32
The settings are using i32 and I'm just going to commit to signed ints
for everything.

No real reason other than it's prolly a little easier even if it's a
little less safe (thought you get negative numbers in i32 and
overflows in u32 hard to know which is worse).