Add single-glyph-drawing helpers to Display
This adds the `draw_glyph_abs_px` method that lets us draw a single
glyph at a pixel precision (i.e. not tied to the grid) as well as
`advance_width_px` which lets us investigate the effective width of a
given character (they're currently all the same value because
monospace, but they could differ).