Add the buttons for numpad mouse movement
The buttons will show the numpad layout in the sidebar and will be
clickable. The intent is to have them always visible, respond to touch
on the web and clickable.

This would make the game fully controllable by mouse/touch (even if it
were a little awkward). They would replace the control hints.

Right now, none of that is true. This just lays the buttons out and
highlights them when hovered over.