Disable keyboard hints at start
I'm not sure if the whole keyboard hint concept should be removed or
not. It feels that with the (clickable!) sidebar numpad UI & Help
screen they're no longer needed.

And I never really liked the idea of them obscuring the map.

So let's disable them for now and see what happens. We can delegate
them into the Settings or something later. Or just drop the code
entirely.