Show the destination on hovered sidebar button
When the player hovers over any of the movement numpad buttons in the
sidebar, the location where the player character would move to is
highlighted on the actual game map.

This helps make the connection that the numpad buttons move the player
clearer.

It also lets the player keep their eyes on the player character and
move the mouse "blindly" -- making relative motions until the desired
destination is highlighted.