Track if the player ever picked up a dose
This fixes the "you've never saved a dose for a later fix" message. It
used to check whether the player was carrying a dose at the moment
ending the game. But the message implied whether they ever picked up a
dose.

This sets a boolean on a dose pickup, which means even if the player
uses it later, it'll still be remembered properly.