Fix mouse highlight positioning in glutin
When the window aspect ratio did not match the expected ratio of the
game (i.e. we were letterboxing), the tile position of a mouse pointer
was not being calculated correctly. We were ignoring how the
letterboxing messes the pos up.

Turns out this was actually solved in glium (and possibly SDL), but
not on glutin which is the currently default backend.

So this fixes it by pretty much copying the glium code wholesale \o/.

With this commit, no matter what I did (fullscreen, windowed,
different width/height ratios) the mouse position was handled
properly.