Make sure the expected amount of items is placed
Since we're varying the map density now, the probabilistic approaches
could result in maps with much fewer items (because the item generator
gives up when it finds an occupied tile).

This calculates the expected amount of items and makes sure that
number is actually placed. We can tweak/randomise the number
explicitly later on if desired.