Replace Rand with oorandom
oorandom was written by icefoxen (of ggez fame) as a zero config
seedable PRNG for use in games. It currently (and possibly forever)
uses PCG behind the hood.

It's a tiny library with no dependencies, supporting no-std
environments and unlikely to change much beyond this point. And shit
if that ain't exactly what I want.

Rand might be awesome at some point but it and Glutin/Winit are the
only dependencies that still ship regular code-breaking dependencies.

So this moves away from Rand to oorandom.

We were using some of Rand's functionality that oorandom doesn't have:
select from a list, weighted selection, that sort of thing.

The seed "generation" is now done by looking at the least significant
bits of the current time (since we're already using Chrono).
Old-school-like.

So this doubles down on the `random` module and `Random` type we've
already introduced and instead of just enabling a typedef, it ships a
proper type with the functionality we actually need in the game.

It's a little bit more code (that could use some cleaning up), but it
works and means that if we ever decide to switch to a different crate,
all this is neatly isolated with a clean interface.

I've benchmarked the map generation and it's about a third faster than
the previous code (which is not surprising as that was using a
CSPRNG). But it doesn't really have any significant effect on the
overall speed.

Shit is fast yo.