Update SDL and fix build errors
This adds the necessary settings-related fixes to the sdl backend.
This should fix the travis errors.

The game plays fine, but there is some weirdness in the font size
switching. When set in the settings.toml file, everything seems fine.

But switching a tilesize in the settings dialog seems to do scaling
the opposite way from what I'd expect (switching to a lower size
increases the apparent tilesize and vice-versa). Also, looks like the
window is not resized to match.