Switch to fullscreen on startup if set in settings
If the settings file we load says we should be in fullscreen, make
sure we are. Previously, that setting was effectively ignored.

We're not doing this on window creation (some backends support it,
some don't). Instead, we always start windowed and only switch to
fullscreen in the first frame if requested.

That means we've only need to handle fullscreen in a single place.