Create a cross Dockerfile with curl
The Linux SDL builds are failing because we're using the `bundled`
feature which has switched from using the `reqwest` Rust crate to
`curl`. Which of course isn't in the official cross images we're
using.

All in the name of reducing dependencies \o/.

Well, anyway I've been meaning to look into what it would mean to add
extra deps to our build images.

So this adds the Dockerfile for `x86_64-unknown-linux-musl` (which we
use in Travis).

I'm actually a little confused because we're building the
`x86_64-unknown-linux-gnu` targets but using the musl cross images? So
this might fail and we might have to build the gnu one instead IDK.

Anyways, the image is at: `docker.io/tsedovic/cross:x86_64-unknown-linux-musl`