Add the gnu cross Dockerfile and override too
Looks like my confusion was based in reality: the actual building of
the Rust code is done in the `x86_64-unknown-linux-gnu` image. For
whatever reason, cross itself runs/is installed in musl.

So: I probably don't need the musl override/image and I'll try to
remove it if this fixes the Linux Ci.