Start moving some of the glutin state code away
This adds a new mod called `loop_state`. It contains some of the
state and data that has been stored in the glutin `main_loop` without
actually being backend-specific.

We've got a ton of code like that which is duplicated in every
backend. So we're trying to tie it up a little and see how much we can
move away and reuse in different backends.