Move some more stuff to the loop_state
I think this is most of it. The full-screen handling is pretty gnarly
and could also use some common code, but unfortunately it's supposed
to trigger the switch from FS to windowed. I guess we could write a
function that returns the "should switch" data in an enum?

That might actually be a good idea.