Fix SDL tilesize changing
We were missing the code to actually resize the window when the
tilesize setting is changed.

Now we've added it. It's a little different from glutin, because
glutin fires off the "window resized" event which makes sure our
display info etc. stays correct.

SDL does not so we need to call
`loop_state.handle_window_size_changed` explicitly upon resizing it.