Move window.set_decorations deeper and disable it
We were calling `set_decorations` right at the beginning of the
`SwitchToFullscreen` match arm. Before we even had a current monitor
available.

Instead, we've now moved it to right before the `set_fullscreen` call.
But also, we've disabled it for now. Seems to not do anything on Linux
and maybe it's not necessary for Windows any more either?