Remove the post-fullscreen force window resizing
I think this is what's messing up with the Windows start-in-fullscreen
behaviour.

At the end of the frame, we ran code that checked the expected window
size and tried to resize the window if it didn't match reality.

But: in fullscreen the expected window size never matches the real
window size because our aspect ratio is different. So this always
called `window.set_inner_size`. And that in turn probably messed up
the situation there.

We should just not do that under fullscreen, but it doesn't seem the
code was doing anything useful any more so I'm just deleting it for
now.