Move CI extra features to Cargo.toml
We were passing "extra features" (such as the SDL static link) in the
CI build scripts. That means we've got two places where we define the
final features for a release build or CI run: Cargo.toml and the
travis/appveyor scripts.

Instead, we've now added a platform-extra-features feature to
Cargo.toml and each build invocation will pass the right one to cargo.

That means if we want to say add an extra feature just to Windows, we
can do so completely in Cargo.toml. It also means we can reproduce the
production builds completely within Cargo as well.

This should also fix the macOS CI because we're no longer requiring
SDL in the `sdl-static-link` feature which is enabled in macOS
unconditionally.

The SDL macOS build is still broken but now it's no longer enabled by
default. We'll figure that out later on (or drop SDL for good).