Remove `padding` from Display
It now only serves the purpose of keeping around a thin line of tiles
so they can be shown partially when scrolling. There's no reason for
the caller to set this.

It needn't even really be exposed (though it kind of is via the
iterators and Point accessors so we'll make the method public).