Resize the Display if the window size changes
The Display should always correspond to what's actually shown on the
window. So when we resize the window (meaning change the number of
tiles we can show), Display should resize with it.

This can be a relatively expensive operation (creating and later on
filling an the full array of displayed tiles). We might need to
optimise it later.

On the other hand, the resizing should be infrequent.