Update the `state.display_size` value too
This fixes the menu centering and sidebar once again takes the full
hight of the window.

But really, we shouldn't even be using that variable in the first
place. Display should be the authoritative value for this. And
actually, possibly for the other stuff like map size too? IDK about
that.