Unify the cfg replay feature code
This drops the dummy functions that don't do anything but allow us to
disable the features.

These functions can have complex signatures that we then have to
update in more places than necessary.

So instead, the "real" and "empty" implementations live in the
single function's body instead. That means we get conditional
compilation, but only a single fn header.