Limit the wasm max canvas size
While we now can make the game area as large as we want, even running
it on full HD kinda tanks the framerate and makes for a worse
experience.

So the previously hardcoded size is now the max but the game can be
shrunk to a smaller size (and therefore fit on a mobile screen).