Allow overriding the default tilesize in wasm
At least for now, the wasm has no settings persistence, so we might
conceivably want to go for a different tilesize (more suited for the
web, especially mobile).

For now, this doens't make any actual changes, but we are now able to
configure `crate::engine::wasm::DEFAULT_TILESIZE` without affecting
any of the other backends.

Ideally, this should be all controlled via the `Settings` struct, even
if it's not exposed to any user modification. But for now, this will
do.