Fix the vertical issues with the sidebar
We're now able to support windows that are 19 tiles tall (this
corresponds the height my phone gets running this in the Firefox Focus
browser in landscape).

Some changes:

* Empty lines get removed when there's no space
* The "Panic" and "Stun" status effects get combined to a single line
* The "numpad controls", "inventory" and "active bonus" labels get hidden
* The actual numpad clickable areas can overlap non-clickable text

That last one isn't awesome, but it shouldn't break functionality and
I can't thin of anything better to do without making even more drastic
changes to the UI.

This should make the game work in browsers with little chrome, but
e.g. in normal Firefox (as opposed to Focus) it's still not enough.

Not sure what to do there other than supporting even smaller tile
size (which I'd rather not -- it's pretty tiny even with the 24px
one).