Fix the wasm mouse position issues
The mouse on the web now works again. The position discrepancies were
caused by the the JS and WASM codebases having different tilesize
values.

Really, we should make sure there's a single source of truth for both
but for now the fix is to have them be the same value.