Remove all tilesize knowledge from the JS code
The JavaScript portion of the codebase doesn't have any need to know
about the tilesize or to be sending the tile positions out.

So now all that info is removed (and therefore there's no longer any
need to sync it between the Rust and JS files) and the JS sends only
the mouse pixel position information and the button status.

Ideally, we'll move even more code to Rust.