Handle DPI correctly
The glutin backend on macOS with a "retina" display now renders the
game in the whole window (as opposed to doing it in a quarter of the
area, leaving the rest empty).

Hte result is blurry, because we're rendering to the expected
"logical" area and then resize it by the DPI factor, which leads to
stretching of the textures.

We'll fix that in another commit.