Fix the tile size changing issues in settings
When a font size was changed in the Settings dialog, the window would
stay the same (good) but, the contents would keep the same number of
tiles, resulting in a display that's either too small or than did not
fit into the window (bad).

Now we not only change the tilesize but also update the display's tile
count.