Use a new Graphic enum intead of glyph char
All of the visuals so far were handled via a `glyph` char. This does
not work when a single conceptual tile (e.g. a tree) can have
different graphical representations (i.e. different graphical tiles).

So this adds a new enum called Graphics and translates it to a
graphical tile or a glyph as needed.