Rename `font.png` to `text.png`
This prepares us for the addition of a separate tileset for the
glyph (ASCII) tilemap.

The idea is to decouple the tiles used for rendering the map (even in
the text version) from the ones used for the text.

We will end up with three tilesets:
1. Text (used for text only, won't go above say ~30 px)
2. Glyphs (used for text-based game tiles, multiples of ten)
3. Tiles (used for the graphics-based game tiles, multiples of ten)

These sizes will therefore be orthogonal to each other and the
settings screen will expose both.

This commit prepares for this by renaming `font.png` to `text.png` and
updating the build script by renaming a few things, etc.