Separate text and tile sizes
This is a fairly large change that makes the text and tile sizes
separate.

Motivation: we might want to render some reasonably big
tiles (30-40px), but the text size can reasonably be 18-20px for
example. Either way, there's no longer any good reason to tie the text
size to the tile size a-priori.

And making say 30px tiles means rendering all the glyphs at 30px,
resulting in a huge tileset we're not going to use for the actual
gameplay.

So: this splits the text and tile sizes and we will (later on) create
an extra texture map for the ASCII-based tiles. Right now, they're
taken from the textmap which doesn't have the right size so it looks
bad.

Remaining work:
* Hook both sizes to the Settings screen
* Make the UI work with the new text / tile size differences
* Generate a separate ASCII tileset
* Hook the ASCII tileset to opengl