Generate a separate glyph tilemap
This is the last piece of the puzzle. We now have three separate
textures: one for printing text, one for ASCII tiles and one for
graphical tiles.

The text size is independent of the tile one.

There are still bugs when it comes to text positioning, but the
ASCII/GFX rendering now works well.