Update the dose, voices and shadow tiles
Instead of the "grainy sparkly" dose switch to one that's visually
more in line with the other ones there. They all now follow a similar
pattern but are hopefully distinctive enough.

I've also replaced the Shadow tile with the former Voices tile and
added a different tile to Voices. They're similar enough to highlight
at a connection as well.

I loved the ghost tiles, but they didn't fit with the other enemies
visually.