Rework some of the colours
We now use green or brown depending on the background being
used (leaves or dirt) and we also update the food colours (to darkish
orange) and the tree options to be more visually distinct.

This has the nice benefit of a) looking PRETTY, b) making it clear
where you can step and where you can't and c) making the shadows,
voices and hunger pop out more.