Use multiple animals tiles when High
We now use four different animal tiles when the player is high. The
tiles always cycle the same way which will let the player grasp how
far on the high scale they are at a glance.

Easier to see than the progress bar (which is still there though).

It also feels wonky and random-y which I like.