Make the starting window size fixed
The game is now able to handle different window sizes so keeping the
window to a "preferred" size/ratio is no longer that important.

And changing the window size based on the tilesize feels more and more
weird the more I actually.

So now the window size is set, determined empirically as the window
size it would have been on default settings.

I plan to disable automatic resizing on tilesize change next as well
as saving the current window size to settings so any changes the
player makes persist.