Add Egui proof of concept
This renders a label and a "Quit" button on top of the game.

The font is incorrect because Egui is hardcoding its fonts currently.
And we're applying no styling to the UI objects.

I'm actually not even sure how much styling we can do -- that's a
thing we'll need to investigate and potentially fix.

Right now, the GUI setup is done in the glutin backend directly. We
should pull that out, expose egui to the game code and let it build
whatever UI it wants.