Consolidate update and render loops
While the code still lives in two separate functions, the code that
processes the window and the code that renders it are now all
happening inside the same window list processing loop right next to
each other.

This is necessary because the "processing" code now "renders" the GUI.
I think ultimately, we'll want to bring it all up to a single
function, but just having them inside the same loop is good enough for
now.