Convert Settings to egui
We're doing some more complex layouts here and also: the settings
window could potentially fit only a part of the screen which means
some of the contents below can be visible (but should not be active).

This required some changes to the window stack loop. I suspect we'll
have to revisit that area some more once we've convert more of the
stuff.