Add egui clip rectangles
The way egui implements hiding the overflow of window contents and the
like is by adding a "clip rect" to each batch of indices/triangles.
The engine is only supposed to draw pixels that are *inside* that
rectangle.

In order to implement this, we had to split the triangle drawing into
batches and adding a clip rect support to the fragment shader.

Luckily, gl::DrawArrays supports passing a subrange of the whole
vertex buffer and we can update the "clip rect" uniform between each
call, so it's not too bad.