Use fixed pos/size for the message box
Turns out, egui is currently crashing when trying to resize a window.
Probably a result of my messing with the codebase's resize memory.
However, it is using a naked unwrap so there.

Anyway, setting the fixed pos/size prevents the window from being
resized and therefore the crashes no longer appear.