Remove extra_px from the engine
This fixes the rendering of the Main Menu borders as well as the
positioning of the version label in the bottom-right corner.

The OpenGL engine now only works with the window pixels, ignoring the
"display size" (i.e. the tile-based area whose size might not match
that exactly).

Mouse and everything else deals with window pixels too.

Of course all the tile-based rendering (which is now only the game and
the sidebar) still use the Display struct and the tile-based
positions, but that works just fine. The only thing to note there is
that the sidebar will currently have black bars at the edges if the
window size doesn't map cleanly to tiles, but we can fix that in later
commits (probably by disentangling the sidebar from the tile positions
and just letting the game's tiles take up all the remaning space).