Add tooltips to Challenge checkboxes in Settings
Each checkboxs now describes what it does (well, would do -- they're
not doing anything yet) on hover.

We had to change the order of the Settings columns because the
tooltips have no overflow detection. If they'd go over the width of
the window, they'll just be cut off instead of being repositioned.