Align the Help, Menu and Movement buttons
Turns out, laying things out at the bottom is a `reverse()` call away.
Sadly, it doesn't seem to work with the `columns` thing we're using
for the button layouts.

So we're positioning that one ourselves for now. The layout's position
from the bottom is hardcoded and that's not great. I think I should be
use some of the calculated heights and use that later on.