Use screen size not display size in egu windows
The terminology is a bit messed up but: to position the in-game windows, we
were calculating the pixel size of the game's outer window by
multiplyong the "display size" bi the tile size. But that's not
nexessarily the same as the actual size of the game window in pixels.

Because the display size is in tiles and the window size might not
match perfectly to tiles.

So this just uses the actual window size (callse "screen size" in the
code) instead. That valuse should always match the game window.