Fix the coordinates that trigger screen recenter
With the sidebar now being done in egui and using a pixel-based value,
the previous logic recentered the screen too early (before the player
could reach the right end of it).

So we now calculate the `map_size` (which recentering code uses to
decide) based on the actual sidebar size.