Fix the incorrect partial rendering during scroll
When the player character moved close to the left or top edge of the
screen, the new tiles being filled in were not rendered properly.
There were some weird rotating glitches in there.

We had some code that converts the vertex-pairs that render a full
tile into smaller ones that only render a part of it. That is busted
for some reason.

BUT: why do we even need that? Especially now that we've added
Scissors and there are no black bars padding the screen.

So I've removed that bit of code and the scrolling now looks great.