Make the sidebar fit on all font sizes
This makes sure that the default window size (1024x768) and higher
will always fit the sidebar for the currently available font
sizes (16, 21 and 28 pixels).

The content is laid out such that none of the buttons will move as the
game progresses (e.g. the player picks up items, gets stunned or has a
companion).

The sidebar is a bit too wide for the smaller font sizes but we'll
address that in a separate commit. Same for the numpad buttons being a
bit too spread out horizontally.