Enforce Help window having always the same height
For some reason, the combination of Window and ScrollArea shrinks if
the contents are smaller than the window (overflow is fine due to the
scroll area). That moves the bottom button around and that's just not
good UX.

Calling `ui.set_min_height` inside the `ScrollArea` fixes this
behaviour by forcing a minimal height too.

Unfortunately, that results in the scrollbar always being shown, but
that can be warked around by setting the min height somewhat smaller
than that of the window.