Make Victory NPC colours variable
Every Victory NPC will now generate a color from a pool just like the
player does. The colour pool is the same as that of the player, but we
try to make sure they never get the same colour.

That way, it should always be obvious who's the player and also
indicate that this is indeed a different person.