Change the license from GPLv3+ to AGPLv3+
Back when Dose Response got started, I had no idea it would ever be
feasible to make the game run inside the browser. As such, the GNU
Affero General Public License seemed unnecessary.

With the advent of WebAssembly, the providing browser support became
viable and indeed a browser-based version of the game is running on
its website.

As such, this commit does what I should have done from the very
beginning: changing the license to AGPL so that the end-users of the
web-based releases benefit from the same fundamental rights (access to
the source code, ability to inspect, modify and publish it under the
same terms) as the desktop users.

If you read this far, please treat all the source code prior to this
relicensing as being under the AGPL as well. I have no legal standing
to demand it, but for what it's worth, it is what I would (and
should!) have done since the very beginning.