Implement the "Fast Depression" Settings option
This is a doozy. Since the Depression speed it's all connected to
creating monsters and therefore map generation but in a way that we
need to tie to the settings now, this touched a lot of code.

We've added a new `Challenge` struct which will carry all the player
challenges. It can be generated from the `Settings` but it is stored
in the `State`.

That means that an active challenge will be persisted during
saving/loading.

The challenge is only refreshed during the start of a new game, so
actually changing the "fast depression" challenge will only appear in
the new game. I think that's more consistent -- you opt into a
challenge, you'll need to keep it.