Use the Challenge system for Hide Unseen Tiles
The renderer is no longer looking at the Settings struct directly to
see whether the "Hide Unseen Tiles" challenge is on or off. Instead,
it uses the `Challenge` struct from `State`. This makes it consistent
with the other challenge, but it means that any change will only
happen in a new game.