Implement the initial Accessible palette
It's not exactly beautiful but it seems to be working well enough.
I've tried it with an android app that simulates a bunch of common colour
vision scenarios and everything seems to be distinguishable.

So I think the colours are fine, but maybe they can be switched around
some. I think the UI bits could definitely use some more touch.