Only generate world when the game is started
Previously, we've generated a world immediately on startup and then
made it so that "new game" actually meant "resume the existed
pregenerated game".

This had issues with the new challenge settings. Since Challenges are
tied to a given World, if you started Dose Response, changed a
Challenge in the Settings and then pressed "New Game", the challenge
wouldn't take effect.

Because the world would have been already generated by that point.

Now the world is only generated when you actually press the equivalent
of "New Game" and it will use whatever Challenge is set at that point.

This needs a bit of care because the "default" world is not actually
playable and it would crash, but I think (hope!) I've captured all the
ways this could happen and in this instance, it's (hopefully!) all
controlled enough that I'm comfortable to have covered them all.