Generate glyphs of the text sizes too
The glyph texture map now contains glyphs not only for all the tile
sizes, but also for the text sizes.

This will be useful for when we want to mix the glyphs and text in the
UI (e.g. the inventory sidebar). While the tiles seem to scale fine,
the glyphs scale really poorly and so it's better to build them at the
expected size.

This could have been done using the text/font texture map. But I think
that one's actually no longer needed for anything now that we use Egui
for all text rendering. So it'll go away soon.