Centre the ASCII glyphs in the texture at build
Since the glyphmap is a texture that's only intended to be used as a
tilemap, rather than for regular text rendering, the glyphs should be
centered there just as they are in the graphical tile texture.

This simplifies the game's rendering code because it doesn't have to
do special positioning based on the texture type.

The `advance_width` map and lookup function is no longer necessary and
so it was removed.