Make the player Stunned/Panicking for AP not Turns
If the player is stunned e.g. three times, it will take three action
points to shrug the effect of, not three turns.

Most of the time, this is a meaningless difference, but it is
important with the "double speed" companion bonus. There, the player
woul have been seemingly stunned for six actions, not three.

This is less confusing.