Mix the sound effects together
Instead of adding all the sounds to the empty queues (and risking
running out of Sinks), let's mix all the sound effects that were
triggered during one frame into a single sound playing on a single
sink.

The way we do this is kind of super messy and I'd like to simplify it,
but it works. We hear multiple distinct sound effects and we're no
longer running out of sinks.