Package the game's release before uploading
This updates the GitHub Actions scripts and process-artefacts.py to do
all the post-processing that we were doing in the Travis and
AppVeyor's scripts and anything else necessary to package all the bits
in.

Creating a new release now means including the game icons, readme,
version file, license, packaging it all up and uploading that single
file to S3.