Simplify the CI actions scripts
This adds a new script called `release.py` which takes the built
binary, packages everything up for the release and then publishes it
to S3.

These actions are now `ci.yml` which means that every CI run will now
produce a nightly (pull requests don't but I'm okay with that for
now).

This lets me always test the latest releases as well make tweaks to
the CI scripts and have them validated during the PR process rather
than creating a million tags as we've done in the past.

I've considered creating a Rust script to do this instead, but I've
had a lot of the bones in `process-artefacts.py` already so I just
stayed with Python.