Improve AI pathfinding
This tweaks the monster pathfinding such that monsters are treated
walkable but with higher cost.

Which means that if you've got two monsters at the same discance from
the player, they'll still look for empty paths. But if the monsters
are e.g. in the tunnel, they would go down the same path.

This is an improvement over the current situation where the closer
monster would treat the player normally, but the one farther away
would recalculate the path and started moving away from the player.

It's also coupled with a change in the ordering of when each monster
gets processed. The ones closer to the player go first. This means
that a monster can move and then open a space for the one behind.

I've played a few games and it seems to provide much better monster
behaviour. I'm sure there are other possible improvements, but this is
pretty good.