Reduce the number of Sinks for effects to two
This resolves most of the audio issues on MacOS. You still get the
sound distorted and sort of echoing for far longer than it should, but
much less frequently now.

I think the downside is potentially losing some sounds, but that's
probably okay. When I played it, I hadn't really noticed any issues.