Add the window and tray icons
The game should now show its icon in the bar on top of the window, in
the window list bar and in the app-switching modes, where appropriate
on the platform.

Unfortunately, this doesn't do anything on the MacOS -- I think there
the icon is set by how the app is packaged or something, IDK.

But Linux and Windows should now be covered.