Implement proper player mouse movement
When the mouse pointer is in a visible, movable area, it will show a
path from the player to the destination.

While the mouse is held down, player will walk that path.

The pathfinding is ignoring dose radiuses so it's a bit dangerous. You
still want to be really careful about it but I think this actually
works quite well.

You can just keep clicking instead if you'd like.