Stop player running into doses
The TileContents enum now contains the `Irresistible` field which
denotes that the tile is within an irresistible radius of a dose.

The player's cost now take this into an account and will try to route
around irresistible areas so the player doesn't walk into one by
accident when using the mouse.

The code is quite convoluted and slow, but it does seem to work
nicely.