Skip a job when it uses an unknown secret
The pull requests coming from the people who aren't the repo's
collaborators don't have access to the secrets we use to upload
binaries to the artefact storage.

But we don't want the CI to fail in that case. Instead, we want it run
the build checks and just not package and upload the artefacts.