Update the release tarball/zip directory structure
Previously, the tar.gz archives produced by the release.py script
had a directory called `.` at the top level of the archive.

And the zip had no enclosing directory, instead having all the files
in the root archive.

Instead, each archive should have the same format: a single directory
with the game name and version and all the files under it.