Fix the panic on macOS
After the latest dependency updates, runnint the binary on macOS
resulted in a panic. I think due to a different ordering ot the winit
events, the fonts weren't being loaded at the right time and nor was
`ctx.begin_frame`.

This adds an explicit call to `begin_frame` and set the fonts there.
This fixes the macOS crash as well as making things a bit more robust
in the future (we're less dependent on the event ordering).