Add more granular checks to the CI
This adds a regular cargo check, then a cargo check with the expected
features and finally a check+clippy run that denies all warnings.

This means that significant issues (syntax error, build error, test
error) will surface quickly and that we will keep our code
warning-free going forward.

If the warnings are too problematic, I'll just take them out later but
I'd rather maintain a set of warnings I care about than not having any
checks at all.

But I categorically do not want a blanked "deny(warnings)" statement
in the code itself because it massively slows down development and
refactoring and is a major pain in the arse.