Improve sound playback performance
The sound effects were calling `rodio::Decoder` several times every
frame -- an operation which seems quite time intensive. It is much
more effective to create a Buffered decoder which can be
cloned (unlike the unbuffered variant).

The same issue exists for the background music, but the playback there
happens much less frequently so we'll get to it at a later point.