Add various profiling tools
We haven't used the benchmarking much yet, but this sets up a
framework we can use for any future benchmarking needs.

What has been extremely useful is `cargo flamegraph`. We're calling it
through the `bin/perf.sh` script. We make sure it records its data in
the `target` directory and therefore doesn't trigger recompilation
every time a new file is created.

Note: it is much better to perf-tune a replay at full speed because on
a regular playthrough or even the default replay, the empty frames
will throw things off.