Speed up monster pathfinding
Pathfinding is currently quite expensive, because for every tile
considered, we look up the ones in the surrounding maximum dose
irresistibility radius so we can figure out whether a tile is
irresistible and reflect that in our cost calculation.

Given our world nearest_dose etc. implementation, that check is quite
costly. And it's only ever needed for when we're doing the
mouse-controlled pathfinding. Monsters don't care about the dose
radius and when the player is within the irresistible grasp, the mouse
cost isn't being considered either.

So we now only check it the one case where we're finding a potential
path for the player via the mouse. And we're using the simplest but
messiest implementation of just passing another `bool` to the
pathfinding function.

But this resolves all the frame drops the game has in the release
mode!

There's a ton of other things we can do -- from caching the found
paths, to reusing chunks rather than doing a hashmap lookup every
time, to finding a faster hash implementation to storing/reading the
irresistible value from each cell directly.

Not to mention exposing whether we care about irresistible doses in
the cost "function" itself and IDK maybe combining cost and blockers.

But all of that can happen later if need be. What we've done now puts
each update call comfortably under 5 milliseconds in the release mode.