Use RNG to generate the player graphic and colour
Looks like the function for generating the seed wasn't giving us a
good distribution at the least significant digit. Meaning we were
always ending up with the same graphic.

So now we only use the seed for starting the RNG and use the number
generator (which should have better statistical properties) for
everything else.