Fix the sidebar on large text and full-HD screen
When running in the full-HD (1920x1080 px) resolution and the largest
text size, the numpad section of the sidebar would overlap with the
Help button.

This adds a font size check that fixes that particular issue.

The whole game doesn't render well on the default 1024x768 window size
with the largest text. But it does on the most standard resolution out
there.